% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bds_gsed}
\alias{bds_gsed}
\title{Utility table for conversion bds --> lex_gsed}
\format{
A data frame
\describe{
\item{lex_gsed}{GSED item name}
\item{bds}{BDS number}
\item{bdsr}{BDS number for right}
\item{bdsl}{BDs number for left}
\item{type}{g1:direct, g2: direct-right/left, m1:message,
m2: message-right/left, gr: direct-right, gl: direct-left,
mr:message-right, ml:message-left}
\item{bds_label}{Dutch label}
\item{labelEN}{English label}}
}
\source{
This table is constructed from \code{dscore::bds_edited.csv} by
script \code{data-raw/R/bds_gsed_table.R}
}
\usage{
bds_gsed
}
\description{
Utility table for conversion bds --> lex_gsed
}
\keyword{dataset}
