% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bds_lexicon}
\alias{bds_lexicon}
\title{BDS lexicon}
\format{
A data frame
\describe{
\item{bdsnummer}{BDS number}
\item{description}{Variable label - Dutch}
\item{description_EN}{Variable label - English}
\item{expected}{Expected type of unit}}
}
\usage{
bds_lexicon
}
\description{
This table defines the labels and units expected from selected BDS entries.
Thie table is used to match BDS numbers to their definition and accepted
input.
}
\keyword{dataset}
