% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bds_individual.R
\name{convert_bds_individual}
\alias{convert_bds_individual}
\title{Convert json BDS data for single individual to class individual}
\usage{
convert_bds_individual(txt = NULL, schema = NULL, ...)
}
\arguments{
\item{txt}{a JSON string, URL or file}

\item{schema}{A JSON string, URL or file that selects the JSON validation
schema.}

\item{\dots}{Additional parameter passed down to
\code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models =
  "bsmodel"} for setting the broken stick model, or \code{call =
  as.call(...)} for setting proper reference standards.}
}
\value{
An object of class individual or \code{NULL}
}
\description{
This function takes data from a json source and saves it as a an object
of class individual. The function automatically calculates
standard deviation scores and broken stick conditional means per visit.
The function hard-codes the transformations \code{nlreferences::transform2z()}
and \code{nlreferences::transform2y()}.
}
\examples{
fn <- system.file("extdata", "allegrosultum", "client3.json", package = "jamestest")
# q <- convert_bds_individual(fn)
}
\seealso{
\code{\link[jsonlite]{fromJSON}}
}
\author{
Stef van Buuren 2020
}
