% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bds_target.R
\name{convert_bds_target}
\alias{convert_bds_target}
\title{Convert json BDS data for single individual to a tibble}
\usage{
convert_bds_target(
  txt = NULL,
  schema = NULL,
  append_ddi = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{txt}{a JSON string, URL or file}

\item{schema}{A JSON string, URL or file that selects the JSON validation
schema.}

\item{append_ddi}{Should DDI measures be appended?}

\item{verbose}{Show warnings of missing references}

\item{\dots}{Additional parameter passed down to}
}
\value{
A \code{tbl_df} object with 8 columns and a \code{child} attribute
}
\description{
This function takes data from a json source, calculate Z-scores in the
analysis metric, and saves the data as a tibble.
}
\examples{
fn <- system.file("extdata", "smocc", "Laura_S.json", package = "jamestest")
q <- convert_bds_target(fn)
q
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren 2021
}
