% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_as_bds.R
\name{export_as_bds}
\alias{export_as_bds}
\title{Save donordata format as JSON}
\usage{
export_as_bds(data, ids, path = NULL, names = NULL, indent = NULL, ...)
}
\arguments{
\item{data}{Dataset in donordata format (having child and time elements)}

\item{ids}{Integer vector with the id number for the individuals}

\item{path}{Optional. Path where files should be written. By default, \code{path}
is the working directory. The function creates the path if it doesn't already
exist.}

\item{names}{Optional. Character vector with \code{length(ids)} elements used to
construct the file name. Spaces are replaced by underscores.
If not specified, the function uses the \code{ids} vector to create file names.}

\item{indent}{Optional. Integer. Number of spaces to indent when using
\code{jsonlite::prettify()}. When not specified, the function writes minified
json.}

\item{\dots}{Forwarded to \code{\link[=write_bds]{write_bds()}}}
}
\description{
Converts data in \code{donordata} format to a JSON file per child.
}
\note{
If child birth date is unknown, then 2000-01-01 is used.
Birth date of the mother is back-calculated from \code{agem}, and thus imprecise.
}
\examples{
names <- c("Laura S", "Thomas S")
ids <- as.integer(c(34071, 34072))
\dontrun{
export_as_bds(minidata, ids = ids, names = names, indent = 2)
}
}
\author{
Stef van Buuren 2021
}
