% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_bdsreader.R
\name{init_bdsreader}
\alias{init_bdsreader}
\title{Initialize a bdsreader object}
\usage{
init_bdsreader(psn = NULL, xyz = NULL)
}
\arguments{
\item{psn}{A tibble or data frame containing subject-level information. If \code{NULL}, a template is returned.}

\item{xyz}{A tibble or data frame with long-format measurement data. If \code{NULL}, an empty template is returned.}
}
\value{
A list of class \code{"bdsreader"} with two elements:
\describe{
\item{psn}{A tibble with one row per subject. See details below.}
\item{xyz}{A tibble in long format with age-based measurements. See details below.}
}
}
\description{
Constructs a standardized data object for use with bdsreader-compatible tools.
Returns a list with personal and measurement data. If \code{psn} or \code{xyz} are not provided,
default templates are created.
}
\section{Fields in \code{psn}}{

\describe{
\item{id}{Integer. Unique subject ID. Set to -1 if not provided.
Code id == 0 is the current target child}
\item{name}{Character. Child name or alias.}
\item{dob}{Date. Date of birth of the child.}
\item{dobm}{Date. Date of birth of the mother.}
\item{dobf}{Date. Date of birth of the father.}
\item{mendate}{Date. Date of menarche (BDS 312), if known.}
\item{src}{Character. Data source or study identifier.}
\item{dnr}{Character. Donor or provenance code.}
\item{sex}{Character. Sex of the child (e.g., "male", "female").}
\item{gad}{Numeric. Gestational age in days.}
\item{ga}{Numeric. Gestational age in completed weeks.}
\item{smo}{Numeric. Smoking status during pregnancy (0 or 1).}
\item{bw}{Numeric. Birth weight in grams.}
\item{hgtm}{Numeric. Height of mother in centimeters.}
\item{hgtf}{Numeric. Height of father in centimeters.}
\item{agem}{Numeric. Age of mother at birth in years.}
\item{etn}{Character. Ethnicity code.}
\item{pc4}{Character. Postal code (NL-style 4-digit format).}
\item{blbf}{Integer. Birth land biological father (landcode).}
\item{blbm}{Integer. Birth land biological mother (landcode).}
\item{eduf}{Integer. Educational level of father.}
\item{edum}{Integer. Educational level of mother.}
\item{par}{Integer. Parity (number of previous children).}
}
}

\section{Fields in \code{xyz}}{

\describe{
\item{age}{Numeric. Age at measurement (e.g., in days or months).}
\item{xname}{Character. Name of the x-variable (typically "age").}
\item{yname}{Character. Name of the measurement variable (e.g., "hgt").}
\item{zname}{Character. Name of the z-score variable. (e.g. "hgt_z")}
\item{zref}{Character. Reference standard for z-score (e.g., "nl_2010_hgt_male_").}
\item{x}{Numeric. Value of the x-variable.}
\item{y}{Numeric. Observed measurement value.}
\item{z}{Numeric. Standardized z-score.}
}
}

\examples{
bds <- init_bdsreader()
validate_bdsreader(bds)

}
