% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.target.R
\name{is.target}
\alias{is.target}
\title{Is this object of class \code{target}?}
\usage{
is.target(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical
}
\description{
Is this object of class \code{target}?
}
