% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table34.R
\docType{data}
\name{table34}
\alias{table34}
\title{Table 34: Landcodes}
\format{
A \code{data.frame} with 296 rows and 2 variables:
\describe{
\item{code}{Country code}
\item{country}{Country name}
}
}
\source{
\url{https://publicaties.rvig.nl/Landelijke_tabellen/Landelijke_tabellen_32_t_m_60_excl_tabel_35/Landelijke_Tabellen_32_t_m_60_in_csv_formaat}
}
\usage{
table34
}
\description{
Table 34: Landcodes
}
\keyword{datasets}
