% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_target_donordata.R
\name{target_to_donordata}
\alias{target_to_donordata}
\title{Convert target object to donordata}
\usage{
target_to_donordata(x)
}
\arguments{
\item{x}{An object of class \code{target}.}
}
\value{
A list of two tibbles
}
\description{
This function transforms an object of class \code{target} into a
donordata object, a list with elements named \code{"child"}
and \code{"time"}.
}
\examples{
fn <- system.file("extdata", "allegrosultum", "client3.json", package = "jamestest")
tgt <- convert_bds_target(fn)
target_to_donordata(tgt)
}
\author{
Stef van Buuren 2021
}
