% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondata.R
\name{timedata}
\alias{timedata}
\title{Extract time-level information}
\usage{
timedata(x)
}
\arguments{
\item{x}{A list containing element \code{xyz}}
}
\value{
A tibble with time-level data
}
\description{
Extract time-level information
}
\examples{
fn <- system.file("extdata/bds_v2.0/smocc/Laura_S.json", package = "jamesdemodata")
timedata(read_bds(fn))
}
