% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_json.R
\name{validate_json}
\alias{validate_json}
\title{Validate JSON file}
\usage{
validate_json(txt = NULL, schema = NULL, verbose = TRUE)
}
\arguments{
\item{txt}{A JSON string, URL or file to be compared to the schema.}

\item{schema}{The name of one the built-in schema's.
The default (\code{NULL}) loads \code{"bds_schema.json"}.}

\item{verbose}{include an error message when validation fails.}
}
\value{
A \code{boolean} with optional \code{errors} attribute.
}
\description{
This  function takes data from a JSON source and compares it to a JSON schema
object. A \code{boolean} is returned indicating whether or not the provided
JSON data matches the schema.
}
\author{
Arjan Huizing 2019
}
