% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{Verify validity of incoming JSON data}
\usage{
verify(txt, auto_format = TRUE, format = "1.0", schema = NULL)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{auto_format}{Logical. Should the format be read from the data? Default is \code{TRUE}.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"} and \code{"3.0"}. Formats \code{"1.0"} and
\code{"1.1"} are included for backward compatibility only. Use \code{format = "3.0"}
for new applications.}

\item{schema}{A JSON string, URL or file with the JSON validation
schema.}
}
\value{
A list of processed input data. Side effect: messages are
thrown for input fields that do not conform to the expected
data.
}
\description{
This function checks the json input, runs it against a validation schema, and
checks the input json for further processing.
}
\examples{
txt <- system.file("extdata/bds_v3.0/smocc/Laura_S.json", package = "jamesdemodata")
schema <- system.file("schemas/bds_v3.0.json", package = "bdsreader")
p <- verify(txt, schema = schema)
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren, Arjan Huizing 2021
}
