% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bds.R
\name{write_bds}
\alias{write_bds}
\title{Write target data as bds-formatted JSON file}
\usage{
write_bds(
  x = NULL,
  file = NULL,
  schema = NULL,
  indent = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Tibble with an attribute called \code{person}}

\item{file}{File name. The default (\code{NULL}) does not write to a file.}

\item{schema}{The name of one the built-in schema's.
The default (\code{NULL}) loads \code{"bds_schema_v1.1.json"}.}

\item{indent}{Integer. Number of spaces to indent when using
\code{jsonlite::prettify()}. When not specified, the function writes
minified json.}

\item{verbose}{Logical. Print file message?}

\item{\dots}{Passed down to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\value{
A string with bds-formatted JSON codes, or \code{NULL} for invalid
JSON
}
\description{
This function takes data from a tibble with a person attribute
and saves it into JSON bds format (which can be sent to JAMES).
}
\details{
Functions \code{read_bds()} and \code{write_bds()} are inverse operations.

If the date of birth is not known, the conversion uses the
artificial birth date \verb{01 Jan 2000} to calculate measurement
dates from age.
}
\examples{
fn <- system.file("examples", "Laura_S.json", package = "bdsreader")
tgt <- read_bds(fn, append_ddi = TRUE, schema = "bds_schema_v1.0.json")
js <- write_bds(tgt)
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}
}
\author{
Stef van Buuren 2021
}
