% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bds_target.R
\name{convert_bds_target}
\alias{convert_bds_target}
\title{Convert json BDS data for single individual to target object}
\usage{
convert_bds_target(txt = NULL, schema = NULL, ...)
}
\arguments{
\item{txt}{a JSON string, URL or file}

\item{schema}{A JSON string, URL or file that selects the JSON validation
schema.}

\item{\dots}{Additional parameter passed down to \code{fromJSON(txt, ...)}.}
}
\value{
An object of class \code{target} or \code{NULL}
}
\description{
This function takes data from a json source and saves it as a an object
of class \code{target}.
}
\examples{
fn <- system.file("extdata", "allegrosultum", "client3.json", package = "jamestest")
q <- convert_bds_target(fn)
}
\seealso{
\code{\link[jsonlite]{fromJSON}}
}
\author{
Stef van Buuren 2021
}
