% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dob.R
\name{get_dob}
\alias{get_dob}
\title{Get the date of birth of a child}
\usage{
get_dob(x)
}
\arguments{
\item{x}{A tibble with a \code{child} attribute.}
}
\value{
Object of class \code{Date}. If dob is missing,
the function return dob 01-01-2000.
}
\description{
Get the date of birth of a child
}
