% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondata.R
\name{persondata}
\alias{persondata}
\title{Extract person-level information}
\usage{
persondata(x)
}
\arguments{
\item{x}{A tibble with a \code{person} attribute}
}
\value{
A tibble with person-level data
}
\description{
Extract person-level information
}
\examples{
fn <- system.file("examples", "Laura_S.json", package = "bdsreader")
persondata(read_bds(fn, format = 1))
}
