% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bds.R
\name{read_bds}
\alias{read_bds}
\title{Reads selected BDS data of a person}
\usage{
read_bds(
  txt = NULL,
  schema = "bds_schema.json",
  append_ddi = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{schema}{A JSON string, URL or file that selects the JSON validation
schema. The default corresponds to \code{schema = bds_schema.json} and loads
\code{inst/json/bds_schema.json}. The older \code{schema = bds_schema_str.json}
expects that numeric BDS values are specified as characters.}

\item{append_ddi}{Should DDI measures be appended?}

\item{verbose}{Show warnings of missing references}

\item{\dots}{Additional parameter passed down to}
}
\value{
A tibble with 8 columns with a \code{person} attribute
}
\description{
This function takes data from a json source, validates the contents against
a JSON validation schema, perform checks, calcuates the D-score,
calculates Z-scores and transforms the data as a tibble with a \code{person}
attribute.
}
\examples{
fn <- system.file("examples", "Laura_S2.json", package = "bdsreader")
q <- read_bds(fn)
q
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren 2021
}
