% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bds.R
\name{read_bds}
\alias{read_bds}
\title{Reads selected BDS data of a child}
\usage{
read_bds(txt = NULL, schema = NULL, append_ddi = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{schema}{A JSON string, URL or file that selects the JSON validation
schema.}

\item{append_ddi}{Should DDI measures be appended?}

\item{verbose}{Show warnings of missing references}

\item{\dots}{Additional parameter passed down to}
}
\value{
A \code{tbl_df} object with 8 columns and a \code{child} attribute
}
\description{
This function takes data from a json source, calcuates the D-score,
calculates Z-scores and transforms the data as a tibble.
}
\examples{
fn <- system.file("examples", "Laura_S.json", package = "bdsreader")
q <- read_bds(fn)
q
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren 2021
}
