% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bds.R
\name{read_bds}
\alias{read_bds}
\title{Reads selected BDS data of a person}
\usage{
read_bds(
  txt = NULL,
  version = 2L,
  schema = NULL,
  append_ddi = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{version}{Integer. JSON schema version number. There are currently two
schemas supported, versions \code{1} and \code{2}.}

\item{schema}{A file name (including path) with the JSON validation schema.
The \code{schema} argument overrides \code{version}. The function extracts the
character following the string \code{"_v"} in the base name, and overwrites the
\code{version} argument by the integer representation of the found character.}

\item{append_ddi}{Should DDI measures be appended?}

\item{verbose}{Show verbose output for \code{\link[centile:y2z]{centile::y2z()}}}

\item{\dots}{Passed down to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\value{
A tibble with 8 columns with a \code{person} attribute
}
\description{
This function takes data from a json source, validates the contents against a
JSON validation schema, perform checks, calculates the D-score, calculates
Z-scores and transforms the data as a tibble with a \code{person} attribute.
}
\examples{
fn <- system.file("examples", "Laura_S.json", package = "bdsreader")
q <- read_bds(fn, version = 1)
q
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}, \code{\link[centile:y2z]{centile::y2z()}}
}
\author{
Stef van Buuren 2021
}
