% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_schema.R
\name{set_schema}
\alias{set_schema}
\title{Set the schema from user input}
\usage{
set_schema(format = "1.0", schema = NULL)
}
\arguments{
\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{1.0}, \code{1.1} and \code{2.0}.}

\item{schema}{A file name (including path) with the JSON validation schema.
The \code{schema} argument overrides \code{format}. The function extracts the
version number for the basename, and overwrites the
\code{format} argument by version number.}
}
\value{
A list with components \code{format}, \code{schema} and \code{schema_base}.
}
\description{
Set the schema from user input
}
\examples{
set_schema("2.0")

# for development
set_schema(schema = "mydir/bds_v3.0.json")
}
