% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_schema.R
\name{set_schema}
\alias{set_schema}
\title{Set the schema from user input}
\usage{
set_schema(version = 2L, schema = NULL)
}
\arguments{
\item{version}{Integer. JSON schema version number. There are currently two
schemas supported, versions \code{1} and \code{2}.}

\item{schema}{A file name (including path) with the JSON validation schema.
The \code{schema} argument overrides \code{version}. The function extracts the
character following the string \code{"_v"} in the base name, and overwrites the
\code{version} argument by the integer representation of the found character.}
}
\value{
A list with components \code{version}, \code{schema} and \code{schema_base}.
}
\description{
Set the schema from user input
}
\examples{
set_schema(1)
set_schema(schema = "mydir/myschema_v1.0.json")
}
