% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_schema.R
\name{set_schema}
\alias{set_schema}
\title{Set the schema from user input}
\usage{
set_schema(format = "1.0", schema = NULL)
}
\arguments{
\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{"1.0"}, \code{"1.1"}, \code{"2.0"} and \code{"3.0"}. Formats \code{"1.0"} and
\code{"1.1"} are included for backward compatibility only. Use \code{format = "3.0"}
for new applications.}

\item{schema}{A file name (optionally including the path) with the JSON
validation schema.
The \code{schema} argument overrides \code{format}. The function extracts the
version number for the basename, and overwrites the
\code{format} argument by version number.}
}
\value{
A list with components \code{format} (either \code{1.0}, \code{1.1}, \code{2.0} or \code{3.0}),
\code{schema} (long file name) and \code{schema_base} (base file name).
}
\description{
Set the schema from user input
}
\examples{
set_schema("2.0")
}
