% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table34}
\alias{table34}
\title{Table 34: Landcodes}
\format{
A \code{data.frame} with 296 rows and 2 variables:
\describe{
\item{code}{Country code}
\item{country}{Country name}
}
}
\source{
publicaties.rvig.nl, Landelijke_tabellen
}
\usage{
table34
}
\description{
Table 34: Landcodes
}
\keyword{datasets}
