% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{Verify validity of incoming JSON data}
\usage{
verify(txt, schema, ...)
}
\arguments{
\item{txt}{A JSON string, URL or file}

\item{schema}{A JSON string, URL or file with the JSON validation
schema.}

\item{\dots}{Additional parameter passed down to \code{fromJSON(txt, ...)}.}
}
\value{
A list of processed input data. Side effect: messages are
thrown for input fields that do not conform to the expected
data.
}
\description{
This function checks the json input, runs it against a validation schema, and
checks the input json for further processing.
}
\examples{
txt <- system.file("examples", "Laura_S.json", package = "bdsreader")
schema <- system.file("schemas/bds_v1.0.json", package = "bdsreader")
p <- verify(txt, schema)
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
\author{
Stef van Buuren, Arjan Huizing 2021
}
