% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bds.R
\name{write_bds}
\alias{write_bds}
\title{Write target data as bds-formatted JSON file}
\usage{
write_bds(x = NULL, schema = NULL, ...)
}
\arguments{
\item{x}{Tibble with an attribute called \code{person}}

\item{schema}{The name of one the built-in schema's.
The default (\code{NULL}) loads \code{"bds_schema.json"}.}

\item{\dots}{Additional parameters. Currently ignored.}
}
\value{
A string with bds-formatted JSON codes, or \code{NULL} for invalid
JSON
}
\description{
This function takes data from a tibble with a person attribute
and saves it into JSON bds format (which can be sent to JAMES).
}
\details{
Functions \code{read_bds()} and \code{write_bds()} are inverse operations.

If the date of birth is not known, the conversion uses the
artificial birth date \verb{01 Jan 2000} to calculate measurement
dates from age.
}
\examples{
fn <- system.file("examples", "Laura_S.json", package = "bdsreader")
tgt <- read_bds(fn, append_ddi = TRUE)
js <- write_bds(tgt)
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}
}
\author{
Stef van Buuren 2021
}
