% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bds.R
\name{write_bds}
\alias{write_bds}
\title{Write target data as bds-formatted JSON file}
\usage{
write_bds(
  x = NULL,
  auto_format = TRUE,
  format = "v2.0",
  schema = NULL,
  file = NULL,
  organisation = 0L,
  indent = NULL,
  check = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Tibble with an attribute called \code{person}}

\item{auto_format}{Logical. Should a field \code{Format} be written to the result?
Default is \code{TRUE}.}

\item{format}{String. JSON data schema version number. There are currently
three schemas supported: \code{1.0}, \code{1.1} and \code{2.0}.}

\item{schema}{A file name (optionally including the path) with the JSON
validation schema.
The \code{schema} argument overrides \code{format}. The function extracts the
version number for the basename, and overwrites the
\code{format} argument by version number.}

\item{file}{File name. The default (\code{NULL}) returns the json representation
of the data and does not write to a file.}

\item{organisation}{Integer. Organisation number of the caller. Optional.}

\item{indent}{Integer. Number of spaces to indent when using
\code{jsonlite::prettify()}. When not specified, the function writes
minified json.}

\item{check}{Logical. Should function check json to conform to schema?}

\item{verbose}{Logical. Print file message?}

\item{\dots}{Passed down to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\value{
A string with bds-formatted JSON codes, or \code{NULL} for invalid
JSON
}
\description{
This function takes data from a tibble with a person attribute
and saves it into JSON bds format (which can be sent to JAMES).
}
\details{
Functions \code{read_bds()} and \code{write_bds()} are inverse operations.

If the date of birth is not known, the conversion uses the
artificial birth date \verb{01 Jan 2000} to calculate measurement
dates from age.
}
\examples{
fn <- system.file("extdata/bds_v1.0/smocc/Laura_S.json", package = "jamesdemodata")
tgt <- read_bds(fn, format = "1.0", append_ddi = FALSE)
js1 <- write_bds(tgt, format = "1.0")
js2 <- write_bds(tgt, format = "2.0")
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}
}
\author{
Stef van Buuren 2021
}
